requirements.txt
# LangGraph + LangChain
langgraph>=0.2.0
langgraph-checkpoint-sqlite>=1.0.0
langchain-core>=0.3.0
langchain-openai>=0.2.0

# PDF extraction
pypdf>=4.0.0

# Database
sqlalchemy>=2.0.0
aiosqlite>=0.20.0

# Retry logic
tenacity>=8.2.0

# Environment
python-dotenv>=1.0.0

# HTTP
requests>=2.31.0

# Evaluation / tracing
langsmith>=0.1.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
